remove ts3.1 for node/v11
