+++ types/node/v11/package.json
@@ -2,11 +2,6 @@
     "private": true,
     "types": "index",
     "typesVersions": {
-        "<=3.1": {
-            "*": [
-                "ts3.1/*"
-            ]
-        },
         "<=3.6": {
             "*": [
                 "ts3.6/*"
