--- types/node/v11/ts3.1/tsconfig.json
@@ -1,30 +1,0 @@
-{
-    "files": [
-        "index.d.ts",
-        "node-tests.ts"
-    ],
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es6",
-        "lib": [
-            "es6",
-            "dom"
-        ],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "strictFunctionTypes": true,
-        "baseUrl": "../../../",
-        "typeRoots": [
-            "../../../"
-        ],
-        "paths": {
-            "node": [
-                "node/v11"
-            ]
-        },
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true
-    }
-}
