--- types/node/v11/ts3.1/tslint.json
@@ -1,10 +1,0 @@
-{
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "ban-types": false,
-        "no-bad-reference": false,
-        "no-empty-interface": false,
-        "no-single-declare-module": false,
-        "unified-signatures": false
-    }
-}
