+++ types/node/v11/ts3.6/index.d.ts
@@ -3,6 +3,4 @@
 // Typically type modifiations should be made in base.d.ts instead of here
 
 /// <reference path="base.d.ts" />
-
-// tslint:disable-next-line:no-bad-reference
-/// <reference path="../ts3.1/assert.d.ts" />
+/// <reference path="assert.d.ts" />
