+++ types/node/v11/ts3.6/tslint.json
@@ -1,6 +1,10 @@
 {
     "extends": "dtslint/dt.json",
     "rules": {
-        "no-bad-reference": false
+        "ban-types": false,
+        "no-bad-reference": false,
+        "no-empty-interface": false,
+        "no-single-declare-module": false,
+        "unified-signatures": false
     }
 }
